+++ dataclass/CHANGELOG.md
@@ -1,3 +1,6 @@
+## [0.3.0] - 2019/01/19
+* Add @Collection() annotation to use DeepCollectionEquality from `collection` package to compare collections
+
 ## [0.2.0] - 2019/12/25
 * enable usage of generic types, eg. `Product<T>`
 * **BREAKING CHANGE**: change generated class name from `_Class` to `_$Class`